requirements_flask.txt
Flask==2.3.3
opencv-python-headless==4.8.1.78
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2
Pillow==10.0.1
gunicorn==21.2.0